Move app components to components/app
