+++ generator/svelte/src/lib/components/app/MasjidsList.svelte
@@ -1,0 +1,65 @@
+<script lang="ts">
+	import type { IMasjid } from '$lib/types';
+	import entries from 'lodash/entries';
+	import Divider from '../uikit/Divider.svelte';
+	import { SUBSCRIPTION_SIDEOVER_ID } from '$lib/constants';
+	import { toast } from '$lib/stores/toast';
+
+	export let masjids: [string, IMasjid][];
+	export let masjidsListElement: HTMLUListElement;
+
+	const onMasjidSubscribeClick = (name: string) => {
+		const checkbox = masjidsListElement.querySelector<HTMLInputElement>(`#${name}`);
+
+		if (!checkbox) {
+			toast.error('Something went wrong, please try again later');
+			return;
+		}
+
+		checkbox.checked = true;
+	};
+</script>
+
+<div class="w-full">
+	{#each masjids as [name, { iqamas, jumas }], i}
+		<h2 class="flex items-center justify-between">
+			<span>{name}</span>
+
+			<button on:click={() => onMasjidSubscribeClick(name)}>
+				<label class="btn btn-primary btn-sm drawer-button" for={SUBSCRIPTION_SIDEOVER_ID}>
+					Subscribe
+				</label>
+			</button>
+		</h2>
+
+		<div class="overflow-x-auto px-4">
+			<table class="table table-zebra">
+				<thead>
+					<tr>
+						<th>Iqama</th>
+						<th>Time</th>
+					</tr>
+				</thead>
+				<tbody>
+					{#each entries(iqamas) as [iqama, { time }]}
+						<tr>
+							<td>{iqama}</td>
+							<td>{time}</td>
+						</tr>
+					{/each}
+				</tbody>
+			</table>
+			<h3>Jumas</h3>
+			<ul>
+				{#each jumas as juma}
+					<li>
+						{juma}
+					</li>
+				{/each}
+			</ul>
+		</div>
+		{#if i !== masjids.length - 1}
+			<Divider />
+		{/if}
+	{/each}
+</div>
